Added routes to App.tsx
