+++ src/component/Todo.tsx
@@ -1,46 +1,19 @@
 import React, { useEffect, useState } from "react";
 import { TodoInterface } from "../utils/TodoInterface";
-import AddTodo from "./AddTodo";
 import ListTodoItem from "./ListTodoItem";
-import usePost from "./usePost";
-import { v4 as uuidv4 } from "uuid";
 import { url } from "../utils/TodoApi";
 import "./styles.css";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Header from "./Header";
+
 import useFetch from "./useFetch";
 
 function Todo() {
   const [todos, setTodos] = useState<TodoInterface[]>([]);
   const { loading, error } = useFetch();
-  const { fetchPost } = usePost();
   useEffect(() => {
     fetch(url)
       .then((res) => res.json())
       .then((data) => setTodos(data));
   }, []);
-
-  const addTodo = (title: string, dueDate: string) => {
-    const today = new Date();
-    const enteredDate = new Date(dueDate);
-    const newTodo: TodoInterface = {
-      id: uuidv4(),
-      title,
-      dueDate,
-      completed: false,
-    };
-    if (enteredDate < today) {
-      alert("Invalid due date. Please enter a date that is not in the past.");
-      return;
-    }
-    const flag = fetchPost(newTodo);
-    if (!flag) {
-      alert("Not added");
-    }
-
-    setTodos([...todos, newTodo]);
-    alert("Added successfully");
-  };
 
   const deleteTodo = (id: string) => {
     const filterTodo = todos.filter((todo) => todo.id !== id);
@@ -56,27 +29,13 @@ function Todo() {
 
   console.log({ todos });
   return (
-    <>
-      <BrowserRouter>
-        <Header />
-        <Routes>
-          <Route path="/addtodo" element={<AddTodo addTodo={addTodo} />} />
-
-          <Route
-            path="/"
-            element={
-              <ListTodoItem
-                todos={todos}
-                deleteTodo={deleteTodo}
-                handleCheckbox={handleCheckbox}
-                loading={loading}
-                error={error}
-              />
-            }
-          />
-        </Routes>
-      </BrowserRouter>
-    </>
+    <ListTodoItem
+      todos={todos}
+      deleteTodo={deleteTodo}
+      handleCheckbox={handleCheckbox}
+      loading={loading}
+      error={error}
+    />
   );
 }
